pyproject.toml
[tool.poetry]
name = "server"
version = "0.0.0"
description = "The Micebot core application."
authors = ["MiceBot Collaborators."]
license = "MIT"
repository = "https://github.com/micebot/server.git"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.60.1"
uvicorn = "^0.11.5"
sqlalchemy = "^1.3.18"
alembic = "^1.4.2"
psycopg2 = "^2.8.5"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
pyjwt = "^1.7.1"
python-multipart = "^0.0.5"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
coverage = "^5.1"
ipdb = "^0.13.3"
faker = "^4.1.1"
requests = "^2.24.0"
factory-boy = "^2.12.0"
pylama = "^7.7.1"
freezegun = "^0.3.15"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79